Original code
I came up with a coding challenge for people to use in interviews.  I thought of this when looking at a dart board today.  Here it is:

The rules for scoring in darts are as follows:
1. Each section has a base value between 1 and 20.
2. The outer sections are worth double.
3. The inner sections are worth triple.
4. The outer bulls-eye is worth 25.
5. The inner bulls-eye is worth 50.

Given the range 1-60, generate an array of point values that are impossible to score using only one dart.
Here is my answer.